+++ tests/Fixture/Unify/ConstantDefaultBase.php
@@ -1,0 +1,10 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Rasuvaeff\Understudy\Tests\Fixture\Unify;
+
+abstract class ConstantDefaultBase
+{
+    public const int FROM_PARENT = 9;
+}