+++ tests/Fixture/Unify/StaticUntypedReturnContract.php
@@ -1,0 +1,11 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Rasuvaeff\Understudy\Tests\Fixture\Unify;
+
+interface StaticUntypedReturnContract
+{
+    /** @return mixed */
+    public static function ping();
+}